Implementing the Observer Design Pattern
This commit implements the Observer Design Pattern by introducing the Subscriber interface and the Publisher class to manage a list of observers. The CategoryService and LocationService classes are updated to inherit from Publisher and notify registered observers upon data changes (creation, updates, and deletions). A LoggingObserver is added to log these events for tracking purposes.

Task-number: 11.2